+++ R/chk.R
@@ -15,8 +15,6 @@ chk_populations <- function(x, x_name = NULL) {
   if(is.null(x_name)) x_name <- deparse_backtick_chk(substitute(x))
 
   chk_list(x, x_name = x_name)
-  chk_named(x, x_name = x_name)
-  chk_unique(names(x), x_name = x_name)
 
   x_name <- paste("elements of", x_name)